requirements.txt
asgiref==3.6.0
Django==4.1.5
django_debug_toolbar==3.8.1
PyYAML==6.0
sqlparse==0.4.3
